% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{add_tissue_category_totals}
\alias{add_tissue_category_totals}
\title{Add total rows to a data frame if there are multiple tissue categories.}
\usage{
add_tissue_category_totals(d, tissue_categories)
}
\arguments{
\item{d}{A data frame with columns for Slide ID, Tissue Category and
columns to summarize}

\item{tissue_categories}{The tissue categories of interest, ordered.}
}
\value{
A data frame with tissue category totals and rows in order.
}
\description{
Add total rows to a data frame if there are multiple tissue categories.
}
