% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_gadget_addin.R
\name{addin_10_merge}
\alias{addin_10_merge}
\title{Merge cell seg data files}
\usage{
addin_10_merge()
}
\description{
\code{addin_10_merge} opens a GUI that allows you to select a directory containing
cell seg data files from multiple, individual fields. The selected files
will be merged
to a single file which is saved to the source directory.
This is similar to the function of the inForm Merge tab but
does not include the ability to review and reject individual fields.
}
