% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consolidate_gadget_addin.R
\name{addin_20_consolidate}
\alias{addin_20_consolidate}
\title{Consolidate and summarize cell seg data files}
\usage{
addin_20_consolidate()
}
\description{
\code{addin_20_consolidate} opens a GUI that allows you to select multiple
cell seg data files. The selected files will be consolidated
to a single file which is saved to an output directory
along with summary reports for each file.
}
