% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_app.R
\name{addin_30_analysis_app}
\alias{addin_30_analysis_app}
\title{Run the Analysis app}
\usage{
addin_30_analysis_app()
}
\description{
Runs a Shiny app which will read a consolidated data file and
create aggregated statistics from it.
}
