% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_map_viewer_addin.R
\name{addin_35_spatial_map_viewer}
\alias{addin_35_spatial_map_viewer}
\title{Spatial map viewer addin}
\usage{
addin_35_spatial_map_viewer()
}
\description{
This function starts a GUI which allows you to select a nearest neighbors
file and an inForm export directory. Once selected, you can view fields
in the export directory superimposed with a map of nearest neighbor
relationships.
}
