% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unmixing_quality_addin.R
\name{addin_40_unmixing_quality}
\alias{addin_40_unmixing_quality}
\title{Generate an unmixing quality (crosstalk) report for a selected export folder.}
\usage{
addin_40_unmixing_quality()
}
\description{
\code{addin_40_unmixing_quality} opens a GUI that allows you to select
an export folder containing component data files for singleplex samples.
An unmixing quality report is created for the samples and saved in
the source directory.
}
\details{
The report generator tries to identify the Opal fluor for each source file
by looking at the file name. It recognizes "DAPI", "AF", "Opalnnn",
and "Opal_nnn".
It also recognizes three leading digits as the number of an Opal fluor.
}
