% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/component_levels_addin.R
\name{addin_50_component_levels}
\alias{addin_50_component_levels}
\title{Generate a component levels report for a selected export folder.}
\usage{
addin_50_component_levels()
}
\description{
\code{addin_50_component_levels} opens a GUI that allows you to select
an export folder containing component data files for multiplex samples.
A component levels report is created for the samples and saved in
the export directory.
}
