% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_of_top_and_bottom_cells_addin.R
\name{addin_60_mean_of_top_and_bottom_cells}
\alias{addin_60_mean_of_top_and_bottom_cells}
\title{Generate a "Mean of top and bottom cells" report for a
selected merge data file and configuration file.}
\usage{
addin_60_mean_of_top_and_bottom_cells()
}
\description{
\code{addin_60_mean_of_top_and_bottom_cells} opens a GUI that allows you to select
a merge data file containing results of a manufacturing staining run,
and a configuration file defining the columns of interest.
It then creates an Excel workbook and Word document reporting on the data.
}
