% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/staining_consistency_report_addin.R
\name{addin_70_staining_consistency_report}
\alias{addin_70_staining_consistency_report}
\title{Generate a "Staining consistency" report for a
selected merge data file.}
\usage{
addin_70_staining_consistency_report()
}
\description{
\code{addin_70_staining_consistency_report} opens a GUI that allows you to select
a merge data file containing results of a staining run.
It then creates a report on the data.
}
