% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_viewer_app.R
\name{addin_80_annotation_viewer_app}
\alias{addin_80_annotation_viewer_app}
\title{Run the Annotation Viewer app}
\usage{
addin_80_annotation_viewer_app()
}
\description{
Runs a Shiny app which will read a GeoJSON annotation file and
display it
}
