% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_annotations.R
\name{annotations_summary}
\alias{annotations_summary}
\title{summaries the two annotations counts and their names}
\usage{
annotations_summary(list_of_samples)
}
\arguments{
\item{list_of_samples}{list of all annotations folders}
}
\value{
the raw annotations
}
\description{
summaries the two annotations counts and their names
}
