% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_cell_seg_files.R
\name{batched_merge}
\alias{batched_merge}
\title{Batched merging of inForm output from multiple fields.}
\usage{
batched_merge(base_path)
}
\arguments{
\item{base_path}{Path to a directory containing files to merge. All
eligible files in this directory will be merged.}
}
\description{
\code{batched_merge} merges inForm output from individual fields,
creating "Merge" data files in the each subdirectory.
}
