% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_rois.R
\name{build_roi}
\alias{build_roi}
\title{Build ROIs from multiple pieces}
\usage{
build_roi(rois, roi_names)
}
\arguments{
\item{rois}{A named list of all candidate ROI pieces}

\item{roi_names}{Names of the ROIs to include in the result}
}
\value{
A single \code{sf::sfc} object containing the ROI of interest, or \code{NULL}
if \code{roi_names} is \code{NULL}.
}
\description{
Build ROIs from multiple pieces
}
\keyword{internal}
