% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/component_level_helpers.R
\name{cc_and}
\alias{cc_and}
\title{Concatenate with comma and "and"}
\usage{
cc_and(s)
}
\arguments{
\item{s}{String vector}
}
\value{
A string
}
\description{
Concatenate with comma and "and"
}
