% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{check_params}
\alias{check_params}
\title{Check that all requested phenotypes are defined}
\usage{
check_params(params, phenotypes)
}
\arguments{
\item{params}{A list whose names are phenotype names.}

\item{phenotypes}{A named list of phenotype definitions.}
}
\description{
Check that all requested phenotypes are defined
}
