% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{check_phenotypes}
\alias{check_phenotypes}
\title{Check that all requested phenotypes are defined}
\usage{
check_phenotypes(params, phenotypes)
}
\arguments{
\item{params}{Phenotype names.}

\item{phenotypes}{A named list of phenotype definitions.}
}
\description{
Check that all requested phenotypes are defined
}
