% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{choose_directory}
\alias{choose_directory}
\title{Cross-platform choose directory function.}
\usage{
choose_directory(caption = "Select folder", default = "")
}
\arguments{
\item{caption}{Caption for the choose directory dialog}

\item{default}{Starting directory}
}
\value{
The path to the selected directory, or NA if the user canceled.
}
\description{
Cross-platform choose directory function.
}
