% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_annotations.R
\name{Compare_annot_rounded_areas}
\alias{Compare_annot_rounded_areas}
\title{prepare and compare the areas of two annotations}
\usage{
Compare_annot_rounded_areas(list_of_samples)
}
\arguments{
\item{list_of_samples}{list of annotations folders}
}
\value{
Difference of rounded areas of two annotations
}
\description{
prepare and compare the areas of two annotations
}
