% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_annotations.R
\name{compare_annotation}
\alias{compare_annotation}
\title{Provide summary of two annotations}
\usage{
compare_annotation(path_to_annotations_files, path_to_write, pattern = NULL)
}
\arguments{
\item{path_to_annotations_files}{path to a list of annotation files or a single one}

\item{path_to_write}{where to write the summary of annotations and rounded areas}

\item{pattern}{which pattern your folder has,}
}
\value{
write a data frame with two sheets, summarizing information regarding annotations
}
\description{
Main function to creates/compare rounded_areas and summary of annotations
}
