% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/component_levels_addin.R
\name{component_levels_addin}
\alias{component_levels_addin}
\title{Generate a component levels report for a selected export folder.}
\usage{
component_levels_addin()
}
\description{
\code{component_levels_addin} opens a GUI that allows you to select
an export folder containing component data files for multiplex samples.
A component levels report is created for the samples and saved in
the export directory.
}
