% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/component_level_helpers.R
\name{component_ridge_plot}
\alias{component_ridge_plot}
\title{Create a ridge plot for a single component data file}
\usage{
component_ridge_plot(path, quantiles = c(0.1, 0.99))
}
\arguments{
\item{path}{Full path to a component data file}
}
\value{
A list with plot (a ggplot object) and data (a long data frame
containing sampled component data)
}
\description{
Create a ridge plot for a single component data file
}
