% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_density.R
\name{compute_density}
\alias{compute_density}
\title{Compute cell densities from counts and tissue area}
\usage{
compute_density(counts, summary_path,
  pixels_per_micron = getOption("phenoptr.pixels.per.micron"))
}
\arguments{
\item{counts}{A data frame with columns for \code{Slide ID}, \code{Tissue Category},
and counts, such as the output of \code{\link[=count_phenotypes]{count_phenotypes()}}.}

\item{summary_path}{Path to a cell seg data summary table containing
sample names and tissue categories matching \code{counts}.}

\item{pixels_per_micron}{Conversion factor to microns}
}
\value{
A data table with counts converted to density in cells / mm^2
}
\description{
Compute cell densities from counts and tissue area
}
