% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_density.R
\name{compute_density_from_table}
\alias{compute_density_from_table}
\title{Compute cell densities from counts and tissue area}
\usage{
compute_density_from_table(counts, areas, tissue_categories)
}
\arguments{
\item{counts}{A data frame with columns for \code{Slide ID}, \code{Tissue Category},
and counts, such as the output of \link{count_phenotypes}.}

\item{areas}{A data frame containing tissue areas in \eqn{cells / micron^2} with
sample names and tissue categories matching \code{counts}.}

\item{tissue_categories}{A character vector of tissue category names
of interest.}
}
\value{
A data table with counts converted to density in \eqn{cells / mm^2}.
}
\description{
Compute cell density from a table of cell counts and
a table of tissue areas.
}
\seealso{
Other aggregation functions: \code{\link{compute_density_from_cell_summary}},
  \code{\link{compute_h_score}},
  \code{\link{compute_mean_expression_many}},
  \code{\link{compute_mean_expression}},
  \code{\link{compute_positivity_many}},
  \code{\link{compute_positivity}},
  \code{\link{count_phenotypes}},
  \code{\link{counts_to_percents}}
}
\concept{aggregation functions}
