% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/positivity.R
\name{compute_h_score}
\alias{compute_h_score}
\title{Compute H-score for a single marker aggregated by Slide ID}
\usage{
compute_h_score(csd, measure, tissue_categories, score_path = NULL,
  thresholds = NULL)
}
\arguments{
\item{csd}{Cell seg data to use.}

\item{measure}{The cell seg data column to measure (as a character string)}

\item{tissue_categories}{A character vector of tissue category names
of interest.}

\item{score_path}{Optional path to to a score_data file (may be merged or not).}

\item{thresholds}{Optional three element vector with the threshold values.}
}
\value{
A data frame with one row per Slide ID, showing cell counts and
percents in each bin and the H score
}
\description{
Thresholds may come from a score data file or be directly provided.
}
\seealso{
Other aggregation functions: \code{\link{compute_density_from_cell_summary}},
  \code{\link{compute_density_from_table}},
  \code{\link{compute_mean_expression_many}},
  \code{\link{compute_mean_expression}},
  \code{\link{compute_positivity_many}},
  \code{\link{compute_positivity}},
  \code{\link{count_phenotypes}},
  \code{\link{counts_to_percents}}
}
\concept{aggregation functions}
