% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/positivity.R
\name{compute_h_score_from_score_data}
\alias{compute_h_score_from_score_data}
\title{Compute H-Score based on parameters in a score data file}
\usage{
compute_h_score_from_score_data(
  csd,
  score_path,
  tissue_categories = NULL,
  .by = "Slide ID"
)
}
\arguments{
\item{csd}{Cell seg data to use.}

\item{score_path}{Path to to a score_data file (may be merged or not).}

\item{tissue_categories}{optionally specify tissue categories of interest.
If not given, tissue categories are taken from the score file.}

\item{.by}{Column to aggregate by}
}
\value{
A data frame with one row per Slide ID, showing cell counts and
percents in each bin and the H-Score. See \link{compute_h_score}.
}
\description{
Compute H-Score based on parameters in a score data file
}
\seealso{
Other aggregation functions: 
\code{\link{compute_density_from_cell_summary}()},
\code{\link{compute_density_from_table}()},
\code{\link{compute_h_score}()},
\code{\link{compute_mean_expression_many}()},
\code{\link{compute_mean_expression}()},
\code{\link{compute_positivity_many}()},
\code{\link{compute_positivity}()},
\code{\link{count_phenotypes}()},
\code{\link{counts_to_percents}()}
}
\concept{aggregation functions}
