% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_expression.R
\name{compute_mean_expression}
\alias{compute_mean_expression}
\title{Compute mean expression of top-expressing cells for a single phenotype}
\usage{
compute_mean_expression(csd, phenotype, param, percentile = NULL,
  count = NULL)
}
\arguments{
\item{csd}{Cell seg data to use. This should already have been filtered
for the slides or fields of interest.}

\item{phenotype}{A phenotype selector. This will be passed to
\code{\link[phenoptr]{select_rows}}.}

\item{param}{The parameter (column) to report, as a string.}

\item{percentile}{The percentile cutoff for top-expressing cells. For
example, to measure the top quartile, the percentile is 0.75.}

\item{count}{The number of top expressing cells to use. Only one of
\code{percentile} and \code{count} can be provided.}
}
\value{
A data frame with columns for count and mean.
}
\description{
Compute mean expression of top-expressing cells for a single phenotype
}
