% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/positivity.R
\name{compute_positivity}
\alias{compute_positivity}
\title{Compute positivity of a single phenotype}
\usage{
compute_positivity(csd, phenotype, positivity)
}
\arguments{
\item{csd}{Cell seg data to use. This should already have been filtered
for the slides or fields of interest.}

\item{phenotype}{A phenotype selector. This will be passed to
\code{\link[phenoptr]{select_rows}}.}

\item{positivity}{A one-sided formula giving the positivity expression
to evaluate.}
}
\value{
A data frame with columns for count (total number of cells of
the given phenotype),
positive (count of positive cells)
and fraction (fraction of positive cells).
}
\description{
Compute positivity of a single phenotype
}
\seealso{
Other aggregation functions: \code{\link{compute_density_from_cell_summary}},
  \code{\link{compute_density_from_table}},
  \code{\link{compute_h_score}},
  \code{\link{compute_mean_expression_many}},
  \code{\link{compute_mean_expression}},
  \code{\link{compute_positivity_many}},
  \code{\link{count_phenotypes}},
  \code{\link{counts_to_percents}}
}
\concept{aggregation functions}
