% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/positivity.R
\name{compute_positivity_many}
\alias{compute_positivity_many}
\title{Compute positivity of multiple phenotypes}
\usage{
compute_positivity_many(csd, positivity_pairs)
}
\arguments{
\item{csd}{Cell seg data to use. This should already have been filtered
for the slides or fields of interest.}

\item{positivity_pairs}{A named list of pairs (lists) of phenotype selectors
and positivity expressions.}
}
\value{
A data frame with columns for count,
positive count,
and percent for each element of \code{positivity_pairs}.
}
\description{
Compute positivity of multiple phenotypes
}
\seealso{
Other aggregation functions: \code{\link{compute_density}},
  \code{\link{compute_h_score}},
  \code{\link{compute_mean_expression_many}},
  \code{\link{compute_mean_expression}},
  \code{\link{compute_positivity}},
  \code{\link{count_phenotypes}}
}
\concept{aggregation functions}
