% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consolidate_gadget_addin.R
\name{consolidate_addin}
\alias{consolidate_addin}
\title{Consolidate and summarize cell seg data files}
\usage{
consolidate_addin()
}
\description{
\code{consolidate_addin} opens a GUI that allows you to select multiple
cell seg data files. The selected files will be consolidated
to a single file which is saved to an output directory
along with summary reports for each file.
}
