% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_phenotypes.R
\name{consolidate_and_split_cell_seg_data}
\alias{consolidate_and_split_cell_seg_data}
\title{Consolidate cell seg data files from parallel projects}
\usage{
consolidate_and_split_cell_seg_data(csd_files = NULL, data = NULL)
}
\arguments{
\item{csd_files}{A list or vector of paths to cell seg data files.}

\item{data}{A list of data tables from read_cell_seg_data. Exactly one
of \code{files} or \code{data} should be provided.}
}
\value{
A single data frame containing consolidated data and columns for each
single phenotype.
}
\description{
Consolidate several cell seg data files, each with its own \code{Phenotype} column,
into a single file with separate columns for each phenotype.
}
\details{
The
individual files must all have exactly the same \code{Sample Name} and \code{Cell ID}
columns. \link{split_phenotypes} is called to split the \code{Phenotype} columns.
}
\seealso{
Other consolidation functions: \code{\link{consolidate_and_summarize_cell_seg_data}}
}
\concept{consolidation functions}
