% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_phenotypes.R
\name{consolidate_and_summarize_cell_seg_data}
\alias{consolidate_and_summarize_cell_seg_data}
\title{Consolidate cell seg data files from parallel projects
and create summary reports}
\usage{
consolidate_and_summarize_cell_seg_data(csd_files, output_dir,
  update_progress = NULL)
}
\arguments{
\item{csd_files}{A list or vector of paths to cell seg data files.}

\item{output_dir}{Path to a directory where the results will be saved.}

\item{update_progress}{Callback function which is called with progress.}
}
\value{
A single data frame containing consolidated data and columns for each
single phenotype, invisibly.
}
\description{
Calls \link{consolidate_and_split_cell_seg_data} to consolidate the supplied
cell seg data files. Additionally calls \link{write_summary_report} to
create a summary report for each source file and the consolidated data.
}
\details{
Writes the consolidated data to \code{Consolidated_data.txt} in the output
directory.
}
\seealso{
Other consolidation functions: \code{\link{consolidate_and_split_cell_seg_data}}
}
\concept{consolidation functions}
