% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_helpers.R
\name{count_within_summary}
\alias{count_within_summary}
\title{Summarize "count within" distances}
\usage{
count_within_summary(csd, radii, phenotypes = NULL, categories = NULL)
}
\arguments{
\item{csd}{Cell seg data with \code{Cell X Position},
\code{Cell Y Position}, field name and \code{Phenotype} columns.}

\item{radii}{Vector of radii to search within.}

\item{phenotypes}{Optional list of phenotypes to include. If omitted,
will use \code{unique_phenotypes(csd)}. Counts are computed for all
pairs of phenotypes.}

\item{categories}{Optional list of tissue categories to compute within.}
}
\value{
A data frame with summary statistics for each phenotype pair
in each Slide ID.
}
\description{
Computes summary "count within" statistics
for each Slide ID in \code{csd} and each pair of phenotypes in \code{phenotypes}.
See \code{\link[phenoptr:count_within]{phenoptr::count_within()}} for details of the counts and the
summary calculation.
}
