% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_phenotypes.R
\name{counts_to_percents}
\alias{counts_to_percents}
\title{Convert a count table to fractional percents}
\usage{
counts_to_percents(counts)
}
\arguments{
\item{counts}{A table containing count data. All numeric columns are
assumed to contain count data. The table must contain a
column whose name contains "Total" or "All".}
}
\value{
A table containing percent values as decimal fractions.
}
\description{
Converts a table of counts, such as the output of \link{count_phenotypes},
to a table of fractional percents. Percents are computed row-wise, e.g. per
tissue category.
}
\seealso{
Other aggregation functions: \code{\link{compute_density_from_cell_summary}},
  \code{\link{compute_density_from_table}},
  \code{\link{compute_h_score_from_score_data}},
  \code{\link{compute_h_score}},
  \code{\link{compute_mean_expression_many}},
  \code{\link{compute_mean_expression}},
  \code{\link{compute_positivity_many}},
  \code{\link{compute_positivity}},
  \code{\link{count_phenotypes}}
}
\concept{aggregation functions}
