% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{discrete_colors}
\alias{discrete_colors}
\title{Create a discrete palette with \code{n} colors}
\usage{
discrete_colors(n)
}
\arguments{
\item{n}{The number of colors needed}
}
\value{
A color vector of length \code{n}
}
\description{
Create a discrete palette with \code{n} colors
}
