% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_maps.R
\name{ensure_distance_columns}
\alias{ensure_distance_columns}
\title{Make sure field_data has distance columns for both phenotypes}
\usage{
ensure_distance_columns(field_data, phenos)
}
\arguments{
\item{field_data}{Cell seg data for a (subset of) a single field}

\item{phenos}{Named list of (two) phenotypes}
}
\value{
field_data with \code{Distance to <pheno>} and \code{Cell ID <pheno>}
columns.
}
\description{
Make sure field_data has distance columns for both phenotypes
}
