% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{favicon}
\alias{favicon}
\title{Function to insert an Akoya favicon link.
Call from within \code{tags$head}.}
\usage{
favicon()
}
\description{
Function to insert an Akoya favicon link.
Call from within \code{tags$head}.
}
