% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{field_column}
\alias{field_column}
\title{Get the name of the column that distinguishes fields in a merged cell
seg data file.}
\usage{
field_column(csd)
}
\arguments{
\item{csd}{Cell seg data}
}
\value{
A column name (as a string)
}
\description{
Get the name of the column that distinguishes fields in a merged cell
seg data file.
}
