% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unmixing_error_helpers.R
\name{file_to_fluor}
\alias{file_to_fluor}
\title{Guess a fluor name from the name of a singleplex image.}
\usage{
file_to_fluor(path)
}
\arguments{
\item{path}{A file path}
}
\value{
A fluor name
}
\description{
This only works for DAPI, AF and Opal fluors
}
