% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_rois.R
\name{find_bounding_box}
\alias{find_bounding_box}
\title{Get a bounding box for the given annotations as a polygon ROI}
\usage{
find_bounding_box(annotations, export_dir)
}
\arguments{
\item{annotations}{A list of Annotation IDs of interest}

\item{export_dir}{inForm export directory containing segmentation map
files for the annotations of interest}
}
\value{
A rectangular polygon which encloses all the given annotations.
}
\description{
Get a bounding box for the given annotations as a polygon ROI
}
\keyword{internal}
