% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/component_level_helpers.R
\name{fluor_ridge_plot}
\alias{fluor_ridge_plot}
\title{Create a ridge plot for a single fluor in multiple fields}
\usage{
fluor_ridge_plot(d, quants, clipping, name, fill)
}
\arguments{
\item{d}{Data for the fluor}

\item{quants}{Quantiles for the fluor data}

\item{clipping}{Clipping fractions for the fluor}

\item{name}{Fluor name}

\item{fill}{Fill color for the plot}
}
\description{
Create a ridge plot for a single fluor in multiple fields
}
