% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{function_exists}
\alias{function_exists}
\title{Check if a function is available in a package}
\usage{
function_exists(package, fun)
}
\arguments{
\item{package}{Name of the package}

\item{fun}{Name of the function}
}
\value{
TRUE if the package is installed and contains the function.
}
\description{
Check if a function is available in a package
}
