% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{make_nested}
\alias{make_nested}
\title{Make a nested data frame with one row per Slide ID and Tissue Category.}
\usage{
make_nested(csd)
}
\arguments{
\item{csd}{Cell seg data to use, possibly nested already.}
}
\value{
A nested data frame.
}
\description{
Nested data is easier to work with than grouped data when the processing
is complex.
}
