% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_phenotypes.R
\name{make_unique_names}
\alias{make_unique_names}
\title{Make unique names from a list of paths to cell seg data files}
\usage{
make_unique_names(csd_files)
}
\arguments{
\item{csd_files}{A list or vector of paths to cell seg data files.}
}
\value{
A vector of names.
}
\description{
If the files have unique names, returns the base names without extension.
If the files do not have unique names, returns the directory name
plus the file name without extension.
If those are not unique, appends a sequence number to each name.
}
