% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_of_top_and_bottom_cells.R
\name{mean_of_top_and_bottom_cells}
\alias{mean_of_top_and_bottom_cells}
\title{Compute mean expression levels of the top and bottom expressing cells,
per Slide ID or field. This creates a quality report.}
\usage{
mean_of_top_and_bottom_cells(
  csd_path = NULL,
  expression_cols,
  top_count = 20,
  bottom_percentile = 0.1,
  adjacent_max = 3,
  tissue_categories = NULL,
  .by = "Slide ID",
  out_path = NULL
)
}
\arguments{
\item{csd_path}{Path to a merged cell seg data, or NULL. If NULL, a file
chooser is opened to allow file selection.}

\item{expression_cols}{Vector of column names to report.}

\item{top_count}{The number of top-expressing cells to select for
averaging.}

\item{bottom_percentile}{The cutoff for the bottom percentile cells.}

\item{adjacent_max}{The maximum ratio between expression of adjacent fluors}

\item{tissue_categories}{Tissue categories to include, or NULL for all}

\item{.by}{The column to aggregate by}

\item{out_path}{The path to the output file; if \code{NULL}, a date-stamped file
will be written to the same directory as \code{csd_path}.}
}
\value{
The path to the generated file.
}
\description{
Report the results in an Excel workbook in the same directory
as \code{csd_path}.
}
