% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_of_top_and_bottom_cells.R
\name{mean_of_top_and_bottom_cells_report}
\alias{mean_of_top_and_bottom_cells_report}
\title{Create reports showing mean expression of top and bottom cells}
\usage{
mean_of_top_and_bottom_cells_report(
  merge_file,
  config_file,
  by,
  tissue_categories,
  adjacent_max
)
}
\arguments{
\item{merge_file}{Path to a merged or consolidated cell seg data file}

\item{config_file}{Path to a configuration file containing
column names in the merge file, one name per line.}

\item{by}{Column to aggregate by, e.g. "Slide ID" or "Annotation ID".}

\item{tissue_categories}{Tissue categories to report, or NULL to use all.}

\item{adjacent_max}{The maximum ratio between expression of adjacent fluors.}
}
\value{
None
}
\description{
Create reports showing mean expression of top and bottom cells
}
