% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_phenotypes.R
\name{merge_and_split_cell_seg_data}
\alias{merge_and_split_cell_seg_data}
\title{Merge several cell seg data files, each with their own \code{Phenotype} column,
into a single file with separate columns for each phenotype.}
\usage{
merge_and_split_cell_seg_data(files)
}
\arguments{
\item{files}{A list or vector of file paths}
}
\value{
A single data frame containing merged data and columns for each
phenotype.
}
\description{
Merge several cell seg data files, each with their own \code{Phenotype} column,
into a single file with separate columns for each phenotype.
}
