% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_phenotypes.R
\name{merge_and_summarize_cell_seg_data}
\alias{merge_and_summarize_cell_seg_data}
\title{Merge cell seg data files from parallel projects
and create summary reports}
\usage{
merge_and_summarize_cell_seg_data(csd_files, output_dir)
}
\arguments{
\item{csd_files}{A list or vector of paths to cell seg data files.}

\item{output_dir}{Path to a directory where the results will be saved.}
}
\value{
A single data frame containing merged data and columns for each
single phenotype, invisibly.
}
\description{
Calls \link{merge_and_split_cell_seg_data} to merge the supplied
cell seg data files. Additionally calls \link{write_summary_report} to
create a summary report for each source file and the merged data.
}
\details{
Writes the merged data to \code{Consolidated_data.txt} in the output
directory.
}
\seealso{
Other merge functions: \code{\link{merge_and_split_cell_seg_data}}
}
\concept{merge functions}
