% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_maps.R
\name{nearest_neighbor_map}
\alias{nearest_neighbor_map}
\title{Make a nearest neighbor map for a single field}
\usage{
nearest_neighbor_map(csd, field, export_path, pheno1, pheno2, color1,
  color2, show_as = c("from_to", "to_from", "mutual", "none"),
  dot_size = 3)
}
\arguments{
\item{csd}{Cell seg data with distance columns}

\item{field}{Sample Name or Annotation ID to map}

\item{export_path}{Path to a directory containing composite and component
image files from inForm}

\item{pheno1, pheno2}{Phenotype definitions}

\item{color1, color2}{Colors to draw the phenotype dots}

\item{show_as}{Which nearest neighbors should be shown?}

\item{dot_size}{Size of the dots used to show phenotypes}
}
\value{
A ggplot object
}
\description{
The phenotype definitions may be NA, in which case the base field
will be shown with any available phenotype.
}
\details{
Note: Only single phenotypes are supported.
}
