% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_helpers.R
\name{nearest_neighbor_summary}
\alias{nearest_neighbor_summary}
\title{Summarize nearest neighbor distances}
\usage{
nearest_neighbor_summary(
  csd,
  phenotypes = NULL,
  details_path = NULL,
  .by = "Slide ID",
  extra_cols = NULL
)
}
\arguments{
\item{csd}{Cell seg data with \verb{Cell X Position},
\verb{Cell Y Position}, field name and \code{Phenotype} columns.}

\item{phenotypes}{Optional list of phenotypes to include. If omitted,
will use \code{unique_phenotypes(csd)}.}

\item{details_path}{If present, path to save a tab-separated table with
nearest-neighbor data for each cell.}

\item{.by}{Column to aggregate by}

\item{extra_cols}{The names of extra columns to include in the detailed
results.}
}
\value{
A data frame with summary statistics for each phenotype pair
in each Slide ID.
}
\description{
Computes summary nearest neighbor statistics (mean, median, etc)
for each \code{.by} in \code{csd} and each pair of phenotypes in \code{phenotypes}.
}
