% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{order_by_slide_and_tissue_category}
\alias{order_by_slide_and_tissue_category}
\title{Order a data frame by slide ID and tissue category, putting the
categories in the given order and the
"Total" category in the proper place.}
\usage{
order_by_slide_and_tissue_category(d, tissue_categories,
  .by = "Slide ID")
}
\arguments{
\item{d}{A data frame with \code{.by} and Tissue Category columns}

\item{tissue_categories}{A vector of category names in the desired order}

\item{.by}{First column to sort by}
}
\value{
The input, sorted
}
\description{
Order a data frame by slide ID and tissue category, putting the
categories in the given order and the
"Total" category in the proper place.
}
