% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_annotations.R
\name{order_tags}
\alias{order_tags}
\title{Order tags}
\usage{
order_tags(tag_str)
}
\arguments{
\item{tag_str}{}
}
\value{
ordered tags
}
\description{
this function order tags, it is assumed that, for instance,
'#IncludeInResults #TC' and  '#TC #IncludeInResults' are identical
}
