% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{parse_phenotypes}
\alias{parse_phenotypes}
\title{Parse a vector of phenotype names.}
\usage{
parse_phenotypes(...)
}
\arguments{
\item{...}{Phenotypes to be decoded, optionally with names.}
}
\value{
A named list of phenotype selectors for use with
\code{\link[phenoptr:select_rows]{phenoptr::select_rows()}}.
}
\description{
Parse a vector of phenotype names.
}
\section{Details}{

Each phenotype must be either a single phenotype name (e.g. CD3+ or CD8-)
or two or more names separated by a slash (/) or comma (,). Additionally,
a name without a + or - and containing either "Total" or "All" will be
interpreted as meaning "All cells".
}

\examples{
parse_phenotypes("CD3+", "CD3+/CD8-", "Total Cells", Macrophage="CD68+,CD163+")
}
