% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unmixing_error_helpers.R
\name{process_file}
\alias{process_file}
\title{Read a singleplex image and extract signal information}
\usage{
process_file(path, primary_fluor, percentile = NULL, n_bright = null,
  skip_fraction = 1e-04)
}
\arguments{
\item{path}{Path to a component_data.tif file}

\item{primary_fluor}{Component name of the signal fluor}

\item{percentile}{Percentile for high-expressing pixels. If omitted,
the brightest n_bright pixels will be used after discarding \code{skip_percent}.}

\item{skip_fraction}{The fraction of brightest pixels that will be ignored.}
}
\value{
A list containing
\itemize{
\item signals - The mean expression of each fluor in the selected bright pixels
\item sn - The expression of non-primary fluors as a fraction of the primary
\item primary - Component plane for the primary fluor (a large matrix)
}
}
\description{
Read a singleplex image and extract signal information
}
