% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_rois.R
\name{process_rois}
\alias{process_rois}
\title{Process ROIs for the cells in csd and associated tissue category areas.}
\usage{
process_rois(csd, study_dir, export_dir, output_dir, require_include)
}
\arguments{
\item{csd}{Consolidated cell seg table}

\item{study_dir}{Directory containing annotation files for the source
images in \code{csd}}

\item{export_dir}{A directory containing \code{binary_seg_maps} files and
merged cell seg data summary files for processing.}

\item{output_dir}{Directory where new files will be written.}

\item{require_include}{Should the result
include only cells contained in ROIs tagged with \verb{#IncludeInResults}?}
}
\value{
A list containing three items:
\itemize{
\item \code{csd}: \code{csd} filtered by \verb{#ExcludeFromResults} and (optionally)
\verb{#IncludeInResults} ROIs, and with extra columns added to show
membership in any other tagged ROIs.
\item \code{removed}: A data frame with counts of cells removed.
\item \code{stats}: A data frame with counts of cells in other tags.
}
}
\description{
\code{csd} will be filtered according to ROIs in the associated images'
annotation files.
}
\details{
Updated summary data files with corrected tissue category areas
will be written to \code{output_dir}.
}
\keyword{internal}
