% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_rois.R
\name{process_rois_single}
\alias{process_rois_single}
\title{Process ROIs for a single image, removing cells that don't meet
the inclusion and exclusion criteria.}
\usage{
process_rois_single(
  sample_name,
  data,
  annotation_paths,
  require_include,
  output_dir
)
}
\arguments{
\item{sample_name}{Name of the sample image (without the extension)}

\item{data}{Cell seg data for the sample image (only)}

\item{annotation_paths}{Full paths to all candidate annotation files}

\item{require_include}{Should the result
include only cells contained in ROIs tagged with \verb{#IncludeInResults}?}

\item{output_dir}{Directory to save a check plot}
}
\value{
A list containing these items:
\itemize{
\item \code{csd}: \code{data} filtered by \verb{#ExcludeFromResults} and (optionally)
\verb{#IncludeInResults} ROIs, and with extra columns added to show
membership in any other tagged ROIs.
\item \code{removed}: A data frame with counts of cells removed.
\item \code{stats}: A data frame with counts of cells in other tags.
\item \code{include_roi}: The union of all #Include ROIs, or NULL if none.
\item \code{exclude_roi}: The union of all #Exclude ROIs, or NULL if none.
}
}
\description{
Process ROIs for a single image, removing cells that don't meet
the inclusion and exclusion criteria.
}
\keyword{internal}
