% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_rois.R
\name{process_tissue_categories_single}
\alias{process_tissue_categories_single}
\title{Get corrected tissue category areas for a single image.}
\usage{
process_tissue_categories_single(
  sample_name,
  annotations,
  include_roi,
  exclude_roi,
  export_dir,
  output_dir
)
}
\arguments{
\item{sample_name}{Name of the image being processed}

\item{annotations}{Annotation IDs of interest}

\item{include_roi}{ROI of include area, or \code{NULL} to include all}

\item{exclude_roi}{ROI of exclude area, or \code{NULL} if none}

\item{export_dir}{Path to directory containing \code{binary_seg_maps} files.}

\item{output_dir}{Directory to write a check plot.}
}
\value{
A data frame with columns for annotation ID, tissue category
and tissue category area.
}
\description{
Get corrected tissue category areas for a single image.
}
\keyword{internal}
