% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_rois.R
\name{read_tagged_rois}
\alias{read_tagged_rois}
\title{Read and combine all tagged ROIS from an annotation file}
\usage{
read_tagged_rois(annotation_file)
}
\arguments{
\item{annotation_file}{Path to file}
}
\value{
A named list of polygons, one for each tag in the annotations
}
\description{
Read and combine all tagged ROIS from an annotation file
}
\keyword{internal}
