% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_annotations.R
\name{rounded_area}
\alias{rounded_area}
\title{Round areas, and compute them for two annotations}
\usage{
rounded_area(poly)
}
\arguments{
\item{poly}{annotation list}
}
\value{
sum of areas
}
\description{
all annotations are rounded to 2 decimal points and areas are computed
}
