% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signal_to_noise_addin.R
\name{signal_to_noise_addin}
\alias{signal_to_noise_addin}
\title{Generate a signal-to-noise (crosstalk) report for a selected export folder.}
\usage{
signal_to_noise_addin()
}
\description{
\code{signal_to_noise_addin} opens a GUI that allows you to select
an export folder containing component data files for singleplex samples.
A signal-to-noise report is created for the samples and saved in
the source directory.
}
\details{
The report generator tries to identify the Opal fluor for each source file
by looking at the file name. It recognizes "DAPI", "AF", and "Opal_nnn".
It also recognizes three leading digits as the number of an Opal fluor.
}
