% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reports.R
\name{signal_to_noise_report}
\alias{signal_to_noise_report}
\title{Create a signal-to-noise report for simplex samples}
\usage{
signal_to_noise_report(export_path = NULL)
}
\arguments{
\item{export_path}{Path to a directory containing component_data files.}
}
\description{
Create a signal-to-noise report for simplex samples
}
