% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_map_viewer_addin.R
\name{spatial_map_viewer}
\alias{spatial_map_viewer}
\title{Run the spatial map viewer with the given data
file and export directory.}
\usage{
spatial_map_viewer(csd_path, export_path)
}
\arguments{
\item{csd_path}{Path to a Consolidated_data.txt file from the consolidation
app or nearest_neighbors.csv file created by the analysis app.}

\item{export_path}{Path to a directory containing composite and
component images for the fields in the data file.}
}
\value{
None; starts the viewer app
}
\description{
See \code{\link[=addin_35_spatial_map_viewer]{addin_35_spatial_map_viewer()}}
for a GUI front-end to this function.
}
