% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_map_viewer_addin.R
\name{spatial_map_viewer_front_end}
\alias{spatial_map_viewer_front_end}
\title{Request input file and directory for the viewer, then launch the viewer}
\usage{
spatial_map_viewer_front_end()
}
\description{
Request input file and directory for the viewer, then launch the viewer
}
