% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reports.R
\name{spatial_statistics_report}
\alias{spatial_statistics_report}
\title{Spatial statistics report}
\usage{
spatial_statistics_report(
  csd = NULL,
  csd_path = NULL,
  export_path,
  from_pheno,
  to_pheno,
  output_file = NULL,
  tissue_categories = NULL
)
}
\arguments{
\item{csd}{Consolidated cell seg data (from \code{Consolidated_data.txt}),
or NULL.}

\item{csd_path}{Path to \code{Consolidated_data.txt} or NULL.}

\item{export_path}{Path to a directory containing composite and component
image files from inForm.}

\item{from_pheno, to_pheno}{Phenotype definitions.
Each parameter is a character
vector to be parsed by \code{\link[phenoptr:parse_phenotypes]{phenoptr::parse_phenotypes()}}.}

\item{output_file}{If provided, the full path to the output file.
If omitted, the report will be written to the \code{export_path} directory.}

\item{tissue_categories}{Tissue categories of interest. If supplied, the
generated report will reflect only cells in these categories. Note:
providing this parameter may significantly slow the report generation.}
}
\description{
Create a report showing the G function (\code{\link[spatstat:Gcross]{spatstat::Gcross()}}),
L function (\code{\link[spatstat:Lcross]{spatstat::Lcross()}}) and pair correlation function
(\code{\link[spatstat:pcfcross]{spatstat::pcfcross()}}) for two phenotypes and all fields in an
experiment.
}
