% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_phenotypes.R
\name{split_phenotype_column}
\alias{split_phenotype_column}
\title{Split a single phenotype column}
\usage{
split_phenotype_column(csd, column)
}
\arguments{
\item{csd}{Cell seg data to use.}

\item{column}{The name of the column to split}
}
\value{
A new data frame with the \code{column} column replaced with
individual columns per phenotype and the \code{Confidence} column(s) removed.
}
\description{
A column containing multiple phenotypes is split
into multiple columns, one for each single phenotype.
}
\details{
Multiple phenotypes in the original column must be separated with "/".
The names of positive phenotypes must end in "+".
}
\keyword{internal}
