% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_phenotypes.R
\name{split_phenotypes}
\alias{split_phenotypes}
\title{Split a phenotype column containing multiple positivities
into multiple columns, one for each single phenotype.}
\usage{
split_phenotypes(csd)
}
\arguments{
\item{csd}{Cell seg data to use.}
}
\value{
A new data frame with the \code{Phenotype} column replaced with
individual columns per phenotype and the \code{Confidence} column removed.
}
\description{
Split a phenotype column containing multiple positivities
into multiple columns, one for each single phenotype.
}
