% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/staining_consistency_report_addin.R
\name{staining_consistency_report}
\alias{staining_consistency_report}
\title{Actually generate the staining consistency report}
\usage{
staining_consistency_report(csd_path, marker, compartment, output_dir = NULL)
}
\arguments{
\item{csd_path}{Path to a merge cell seg data file}

\item{marker}{Name of the marker to report}

\item{compartment}{Name of the compartment to report}

\item{output_dir}{(Optional) Directory for the result, default
is the directory containing \code{csd_path}.}
}
\value{
The path to the created file.
}
\description{
Actually generate the staining consistency report
}
