% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_rois.R
\name{summarize_cell_seg_data}
\alias{summarize_cell_seg_data}
\title{Create a summary data file from consolidated cell seg data and, optionally,
tissue areas.}
\usage{
summarize_cell_seg_data(csd, areas = NULL)
}
\arguments{
\item{csd}{Consolidated cell seg data to summarize}

\item{areas}{Data frame of tissue area by annotation and category}
}
\value{
A data frame with columns for Slide ID, Annotation ID,
Tissue Category, cell counts for each positive phenotype, and tissue area
}
\description{
Create a summary data file from consolidated cell seg data and, optionally,
tissue areas.
}
\keyword{internal}
