% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reports.R
\name{unmixing_error_report}
\alias{unmixing_error_report}
\title{Create an unmixing error report for simplex samples}
\usage{
unmixing_error_report(export_path = NULL)
}
\arguments{
\item{export_path}{Path to a directory containing component_data files.}
}
\description{
\code{export_path} should be the path to
an export folder containing component data files for singleplex samples.
An unmixing error report is created for the samples and saved in
the source directory.
}
\details{
The report generator tries to identify the Opal fluor for each source file
by looking at the file name. It recognizes "DAPI", "AF", "Opalnnn"
and "Opal_nnn".
It also recognizes three leading digits as the number of an Opal fluor.
}
