% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_rois.R
\name{update_summary_data}
\alias{update_summary_data}
\title{Find merged cell seg summary files in \code{export_dir}. Make corrected ones by
substituting tissue category areas in the files with the ones in \code{areas}.}
\usage{
update_summary_data(export_dir, output_dir, areas)
}
\arguments{
\item{export_dir}{An inForm export directory containing merged cell seg
data files.}

\item{output_dir}{Directory to receive corrected files.}

\item{areas}{A data frame with columns for Annotation ID, Tissue Category
and category area.}
}
\description{
Write new files with corrected areas and selected columns to \code{output_dir}.
}
\keyword{internal}
