% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upset_plot.R
\name{upset_data}
\alias{upset_data}
\title{Make a data set for use with UpSetR}
\usage{
upset_data(d)
}
\arguments{
\item{d}{A cell seg table with multiple \code{Phenotype} columns}
}
\value{
A data frame with just the \code{Phenotype} columns, coded as 0/1.
}
\description{
\link[UpSetR:upset]{UpSetR::upset} wants sequential columns and 0/1 data.
This function selects the phenotype columns and converts them to 0/1
}
