% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upset_plot.R
\name{upset_plot}
\alias{upset_plot}
\title{Create an UpSet plot showing the phenotype combinations present in data.}
\usage{
upset_plot(csd, expected = NULL)
}
\arguments{
\item{csd}{A cell seg table with multiple \code{Phenotype} columns}

\item{expected}{Optional vector of expected phenotypes. If provided,
these will be highlighted in the plot.}
}
\value{
An upset plot created with \link[UpSetR:upset]{UpSetR::upset}.
}
\description{
See http://caleydo.org/tools/upset/ for an explanation of the plot.
}
