% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel_helpers.R
\name{write_counts_sheet}
\alias{write_counts_sheet}
\title{Write a counts table to an Excel workbook}
\usage{
write_counts_sheet(wb, counts, sheet_name = "Cell Counts",
  sheet_title = "Cell Counts")
}
\arguments{
\item{wb}{An openxlsx Workbook}

\item{counts}{A data frame with columns for \code{Slide ID}, \code{Tissue Category},
and counts, such as the output of \code{\link[=count_phenotypes]{count_phenotypes()}}.}

\item{sheet_name}{Optional name for the worksheet.}

\item{sheet_title}{Optional title header for the table.}
}
\description{
Write a counts table to an Excel workbook
}
