% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel_helpers.R
\name{write_density_sheet}
\alias{write_density_sheet}
\title{Write a density table to an Excel workbook}
\usage{
write_density_sheet(wb, densities, sheet_name = "Cell Densities",
  sheet_title = "Cell Densities (cells/mm2)")
}
\arguments{
\item{wb}{An openxlsx Workbook}

\item{densities}{A data frame with columns for \code{Slide ID}, \code{Tissue Category},
\code{Tissue Area} and densities, such as the output of \code{\link[=compute_density]{compute_density()}}.}

\item{sheet_name}{Optional name for the worksheet.}

\item{sheet_title}{Optional title header for the table.}
}
\description{
Write a density table to an Excel workbook
}
