% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel_helpers.R
\name{write_expression_sheet}
\alias{write_expression_sheet}
\title{Write an expression table to an Excel workbook}
\usage{
write_expression_sheet(wb, exprs, sheet_name = "Mean Expression",
  sheet_title = "Mean Expression")
}
\arguments{
\item{wb}{An openxlsx Workbook}

\item{exprs}{A data frame with columns for \code{Slide ID}, \code{Tissue Category},
and mean expression, such as the output of
\code{\link[=compute_mean_expression_many]{compute_mean_expression_many()}}. Count columns will not be reported.}

\item{sheet_name}{Optional name for the worksheet.}

\item{sheet_title}{Optional title header for the table.}
}
\description{
Write an expression table to an Excel workbook
}
