% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel_helpers.R
\name{write_h_score_sheet}
\alias{write_h_score_sheet}
\title{Write an H-Score table to an Excel workbook}
\usage{
write_h_score_sheet(wb, h_score, sheet_name = "H-Score",
  sheet_title = "H-Score")
}
\arguments{
\item{wb}{An openxlsx Workbook}

\item{h_score}{A data frame with columns for \code{Slide ID}, \code{Tissue Category},
and percent.}

\item{sheet_name}{Optional name for the worksheet.}

\item{sheet_title}{Optional title header for the table.}
}
\description{
Write an H-Score table to an Excel workbook
}
