% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_of_top_and_bottom_cells.R
\name{write_mean_of_top_and_bottom_charts}
\alias{write_mean_of_top_and_bottom_charts}
\title{Create summary charts from the results of \code{mean_of_top_and_bottom_cells}}
\usage{
write_mean_of_top_and_bottom_charts(
  worksheet_path,
  output_path,
  .by = "Slide ID"
)
}
\arguments{
\item{worksheet_path}{Path to an Excel file containing sheets written
by \link{mean_of_top_and_bottom_cells}.}

\item{output_path}{Path to write the resulting file.}

\item{.by}{The aggregation column name}
}
\description{
Create a Microsoft Word file or HTML document containing summary charts
derived from the output of \link{mean_of_top_and_bottom_cells}
with default parameters.
The file type is determined by the file extension
of \code{output_path}, which must be either \code{.docx} or \code{.html}.
}
