% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel_helpers.R
\name{write_percents_sheet}
\alias{write_percents_sheet}
\title{Write a cell percent table to an Excel workbook}
\usage{
write_percents_sheet(wb, percents, sheet_name = "Cell Percents",
  sheet_title = "Percentage of Total Cells")
}
\arguments{
\item{wb}{An openxlsx Workbook}

\item{percents}{A data frame with columns for \code{Slide ID}, \code{Tissue Category},
and percent.}

\item{sheet_name}{Optional name for the worksheet.}

\item{sheet_title}{Optional title header for the table.}
}
\description{
Write a cell percent table to an Excel workbook
}
