% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel_helpers.R
\name{write_plot_sheet}
\alias{write_plot_sheet}
\title{Write a plot to an Excel workbook}
\usage{
write_plot_sheet(
  wb,
  plot,
  sheet_name = "Phenotypes",
  sheet_title = "All combinations of phenotypes in all slides"
)
}
\arguments{
\item{wb}{An openxlsx Workbook from \link[openxlsx:createWorkbook]{openxlsx::createWorkbook}}

\item{plot}{A plot such as the output from \link{upset_plot}.}

\item{sheet_name}{Optional name for the worksheet.}

\item{sheet_title}{Optional title header for the plot.}
}
\description{
Write a plot to a sheet in an Excel workbook.
}
\seealso{
Other output functions: 
\code{\link{write_count_within_sheet}()},
\code{\link{write_counts_sheet}()},
\code{\link{write_density_sheet}()},
\code{\link{write_expression_sheet}()},
\code{\link{write_h_score_sheet}()},
\code{\link{write_nearest_neighbor_summary_sheet}()},
\code{\link{write_percents_sheet}()},
\code{\link{write_summary_sheet}()}
}
\concept{output functions}
