% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_rois.R
\name{write_roi_stats}
\alias{write_roi_stats}
\title{Write a workbook with stats from the ROI trimming}
\usage{
write_roi_stats(roi_results, output_dir)
}
\arguments{
\item{roi_results}{Results from \link{process_rois}}

\item{output_dir}{Directory for the report}
}
\description{
Write a workbook with stats from the ROI trimming
}
\keyword{internal}
