% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reports.R
\name{write_session_info}
\alias{write_session_info}
\title{Write session info to a file}
\usage{
write_session_info(path)
}
\arguments{
\item{path}{Path to the output file}
}
\value{
None
}
\description{
This will write the output of \code{sessioninfo::session_info()}
if it is available, otherwise \code{utils::sessionInfo()} is used.
}
