% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel_helpers.R
\name{write_sheet}
\alias{write_sheet}
\title{Write a single sheet with formatting common to all sheets.}
\usage{
write_sheet(wb, d, sheet_name, sheet_title, header_col)
}
\arguments{
\item{wb}{An openxlsx Workbook from \link[openxlsx:createWorkbook]{openxlsx::createWorkbook}}

\item{d}{A data frame to write.}

\item{sheet_name}{Name for the worksheet.}

\item{sheet_title}{Title header for the plot.}

\item{header_col}{Column number to start the \code{sheet_title}}
}
\description{
\itemize{
\item Bold, centered header
\item Bold column headers
\item Bold Slide ID column
}
}
