% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reports.R
\name{write_summary_charts}
\alias{write_summary_charts}
\title{Create summary charts from the results of an analysis}
\usage{
write_summary_charts(
  workbook_path,
  output_path,
  .by = "Slide ID",
  max_slides_per_plot = 20,
  max_heatmaps_per_plot = 8
)
}
\arguments{
\item{workbook_path}{Path to an Excel file containing sheets written
by \link{write_counts_sheet}, etc.}

\item{output_path}{Path to write the resulting file.}

\item{.by}{Name of the grouping parameter in the worksheets.}

\item{max_slides_per_plot}{Maximum number of slides or samples
to show on each plot.}

\item{max_heatmaps_per_plot}{Maximum number of heatmaps to show
on each plot.}
}
\description{
Creates a Microsoft Word file or HTML report containing summary charts
derived from an analysis. The file type is determined by the file extension
of \code{output_path}, which must be either \code{.docx} or \code{.html}.
}
