% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reports.R
\name{write_summary_report}
\alias{write_summary_report}
\title{Create a summary report for a cell seg data file}
\usage{
write_summary_report(csd_path = NULL, csd = NULL,
  dataset_name = NULL, output_path)
}
\arguments{
\item{csd_path}{Path to cell seg data file, or NULL.}

\item{csd}{Cell seg table, or NULL.}

\item{dataset_name}{Descriptive name of the dataset.}

\item{output_path}{Path to write the resulting HTML file.}
}
\description{
Creates an HTML report summarizing the slides, fields and phenotypes
in a single cell seg data file.
}
