% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_map_viewer_addin.R
\name{addin_35_spatial_map_viewer}
\alias{addin_35_spatial_map_viewer}
\title{Spatial map viewer addin}
\usage{
addin_35_spatial_map_viewer()
}
\description{
This function starts a GUI which allows you to select a
\code{Consolidated_data.txt} file and an inForm image directory.
Once selected, you can view fields from the image directory with a
superimposed map of nearest neighbor relationships.
}
