% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{choose_files}
\alias{choose_files}
\title{Cross-platform choose files function}
\usage{
choose_files(
  caption = "Select files",
  default = "",
  multi = TRUE,
  filters = NULL
)
}
\arguments{
\item{caption}{Caption for the choose directory dialog}

\item{default}{Starting directory}

\item{multi}{Allow multiple files to be selected}

\item{filters}{A two-column matrix of filename filters, or a two-element
vector containing a single filter.}
}
\value{
The path to the selected file(s), or NA if the user canceled.
}
\description{
Cross-platform choose files function
}
