% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reports.R
\name{component_levels_report}
\alias{component_levels_report}
\title{Create a component levels report for multiplex samples}
\usage{
component_levels_report(export_path = NULL)
}
\arguments{
\item{export_path}{Path to a directory containing component_data files.}
}
\description{
\code{export_path} should be the path to
an export folder containing component data files for multiplex samples.
A component levels report is created for the samples and saved in
the source directory.
}
