% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_density.R
\name{compute_density}
\alias{compute_density}
\title{Compute cell densities from counts and tissue area}
\usage{
compute_density(counts, summary_path, tissue_categories,
  pixels_per_micron = getOption("phenoptr.pixels.per.micron"))
}
\arguments{
\item{counts}{A data frame with columns for \code{Slide ID}, \code{Tissue Category},
and counts, such as the output of \link{count_phenotypes}.}

\item{summary_path}{Path(s) to cell seg data summary table(s) containing
sample names and tissue categories matching \code{counts}.}

\item{tissue_categories}{A character vector of tissue category names
of interest.}

\item{pixels_per_micron}{Conversion factor to microns.}
}
\value{
A data table with counts converted to density in  \eqn{cells / mm^2}.
}
\description{
Compute cell density from a table of cell counts and tissue areas
read from a summary cell seg data file.
}
\seealso{
Other aggregation functions: \code{\link{compute_h_score}},
  \code{\link{compute_mean_expression_many}},
  \code{\link{compute_mean_expression}},
  \code{\link{compute_positivity_many}},
  \code{\link{compute_positivity}},
  \code{\link{count_phenotypes}}
}
\concept{aggregation functions}
