% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_expression.R
\name{compute_mean_expression}
\alias{compute_mean_expression}
\title{Compute mean expression of top-expressing cells for a single phenotype}
\usage{
compute_mean_expression(csd, phenotype, param, percentile = NULL,
  count = NULL)
}
\arguments{
\item{csd}{Cell seg data to use. This should already have been filtered
for the slides or fields of interest.}

\item{phenotype}{A phenotype selector. This will be passed to
\link[phenoptr:select_rows]{phenoptr::select_rows}.}

\item{param}{The parameter (column) to report, as a string.}

\item{percentile}{The percentile cutoff for top-expressing cells. For
example, to measure the top quartile, the percentile is 0.75.}

\item{count}{The number of top expressing cells to use. Only one of
\code{percentile} and \code{count} can be provided. If both are omitted,
the mean expression of all cells is returned.}
}
\value{
A data frame with columns for count and mean.
}
\description{
Find the cells with the highest expression of the given parameter
within the given phenotype.
Report the mean expression of the high-expressing cells.
}
\seealso{
Other aggregation functions: \code{\link{compute_density_from_cell_summary}},
  \code{\link{compute_density_from_table}},
  \code{\link{compute_h_score}},
  \code{\link{compute_mean_expression_many}},
  \code{\link{compute_positivity_many}},
  \code{\link{compute_positivity}},
  \code{\link{count_phenotypes}},
  \code{\link{counts_to_percents}}
}
\concept{aggregation functions}
