% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_expression.R
\name{compute_mean_expression_many}
\alias{compute_mean_expression_many}
\title{Compute mean expression of cells for multiple phenotypes and markers.}
\usage{
compute_mean_expression_many(
  csd,
  phenotypes,
  params,
  tissue_categories = NULL,
  percentile = NULL,
  count = NULL,
  .by = "Slide ID"
)
}
\arguments{
\item{csd}{Cell seg data to use. This should already have been filtered
for the slides or fields of interest. It may already be nested by
!!.by and \verb{Tissue Category}.}

\item{phenotypes}{A named list of phenotype selectors
(see \link[phenoptr:parse_phenotypes]{phenoptr::parse_phenotypes}).}

\item{params}{A named list matching phenotype names to
expression column names.}

\item{tissue_categories}{Optional vector of tissue category names to include.}

\item{percentile}{The percentile cutoff for top-expressing cells. For
example, to measure the top quartile, the percentile is 0.75. Negative
numbers will use low-expressing cells; to measure the bottom decile,
use a percentile of -0.1.}

\item{count}{The number of top expressing cells to use. Only one of
\code{percentile} and \code{count} can be provided. If both are omitted,
all cells matching the phenotype are used and the result is the
overall mean expression.}

\item{.by}{Column to aggregate by}
}
\value{
A data frame with a column for mean for each \code{param_pair}.
}
\description{
For each given combination of phenotype and expression parameter,
report the mean expression of the given marker in the specified cells.
}
\details{
This is a very flexible function. If \code{percentile} and \code{count} are both
omitted, it will compute the mean expression for all cells in the
given phenotype. If either \code{percentile} or \code{count} is given, the mean
expression of the highest expressing cells is computed. If a negative
\code{percentile} is given, the lowest percentile is used; for example
\code{percentile=-.1} would give the expression of the lowest-expressing
decile.

By default, this function aggregates by \verb{Slide ID} and \verb{Tissue Category}.
To compute mean expression by a different aggregate, pass a column name in
the \code{.by} parameter.

To aggregate over all cells, include \code{"Total Cells"=NA} as one of the
phenotypes.
}
\seealso{
Other aggregation functions: 
\code{\link{compute_density_from_cell_summary}()},
\code{\link{compute_density_from_table}()},
\code{\link{compute_h_score_from_score_data}()},
\code{\link{compute_h_score}()},
\code{\link{compute_mean_expression}()},
\code{\link{compute_positivity_many}()},
\code{\link{compute_positivity}()},
\code{\link{count_phenotypes}()},
\code{\link{counts_to_percents}()}
}
\concept{aggregation functions}
