% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_expression.R
\name{compute_mean_expression_many}
\alias{compute_mean_expression_many}
\title{Compute mean expression of top-expressing cells for multiple phenotypes}
\usage{
compute_mean_expression_many(csd, phenotypes, params,
  tissue_categories = NULL, percentile = NULL, count = NULL)
}
\arguments{
\item{csd}{Cell seg data to use. This should already have been filtered
for the slides or fields of interest. It may already be nested by
\code{Slide ID} and \code{Tissue Category}.}

\item{phenotypes}{A named list of phenotype selectors
(see \link[phenoptr:parse_phenotypes]{phenoptr::parse_phenotypes}).}

\item{params}{A named list matching phenotype names to
expression column names.}

\item{tissue_categories}{Optional vector of tissue category names to include.}

\item{percentile}{The percentile cutoff for top-expressing cells. For
example, to measure the top quartile, the percentile is 0.75.}

\item{count}{The number of top expressing cells to use. Only one of
\code{percentile} and \code{count} can be provided. If both are omitted,
all cells matching the phenotype are used and the result is the
overall mean expression.}
}
\value{
A data frame with columns for count and mean for each \code{param_pair}.
}
\description{
For each given combination of phenotype and expression parameter,
find the cells with the highest expression within the phenotype.
Report the mean expression of the high-expressing cells.
}
\details{
High expression may be specified by a percentile cutoff or a count of cells.
}
\seealso{
Other aggregation functions: \code{\link{compute_density_from_cell_summary}},
  \code{\link{compute_density_from_table}},
  \code{\link{compute_h_score_from_score_data}},
  \code{\link{compute_h_score}},
  \code{\link{compute_mean_expression}},
  \code{\link{compute_positivity_many}},
  \code{\link{compute_positivity}},
  \code{\link{count_phenotypes}},
  \code{\link{counts_to_percents}}
}
\concept{aggregation functions}
