% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/positivity.R
\name{compute_positivity_many}
\alias{compute_positivity_many}
\title{Compute positivity of multiple phenotypes}
\usage{
compute_positivity_many(csd, phenotypes, positivity_pairs,
  tissue_categories = NULL)
}
\arguments{
\item{csd}{Cell seg data to use. This should already have been filtered
for the slides or fields of interest.}

\item{phenotypes}{A named list of phenotype selectors
(see \link[phenoptr:parse_phenotypes]{phenoptr::parse_phenotypes}).}

\item{positivity_pairs}{A named list of pairs (lists) of phenotype names
and positivity expressions. The expressions must be one-sided formulas
such as \code{~`Membrane PDL1 (Opal 520) Mean`>pdl1_threshold}.}

\item{tissue_categories}{Optional vector of tissue category names to include.}
}
\value{
A data frame with columns for count,
positive count,
and percent for each element of \code{positivity_pairs}.
}
\description{
Compute positivity of multiple phenotypes
}
\seealso{
Other aggregation functions: \code{\link{compute_density_from_cell_summary}},
  \code{\link{compute_density_from_table}},
  \code{\link{compute_h_score}},
  \code{\link{compute_mean_expression_many}},
  \code{\link{compute_mean_expression}},
  \code{\link{compute_positivity}},
  \code{\link{count_phenotypes}}
}
\concept{aggregation functions}
