% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_phenotypes.R
\name{consolidate_and_summarize_cell_seg_data}
\alias{consolidate_and_summarize_cell_seg_data}
\title{Consolidate cell seg data files from parallel projects
and create summary reports}
\usage{
consolidate_and_summarize_cell_seg_data(csd_files, output_dir,
  update_progress = NULL)
}
\arguments{
\item{csd_files}{A list or vector of paths to cell seg data files.}

\item{output_dir}{Path to a directory where the results will be saved.}

\item{update_progress}{Callback function which is called with progress.}
}
\value{
A single data frame containing consolidated data and columns for each
single phenotype, invisibly.
}
\description{
Consolidate several cell seg data files,
each with its own \code{Phenotype} column,
into a single file with separate columns for each phenotype.
}
\details{
Create a summary report for each source file and the consolidated data.

Write the consolidated data to \code{Consolidated_data.txt} in the output
directory.

The
individual files must all have exactly the same
\code{Sample Name} or \code{Annotation ID} and \code{Cell ID}
columns. \link{split_phenotypes} is called to split the \code{Phenotype} columns.
}
