% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_phenotypes.R
\name{consolidate_and_summarize_cell_seg_data}
\alias{consolidate_and_summarize_cell_seg_data}
\title{Consolidate cell seg data files from parallel projects
and create summary reports}
\usage{
consolidate_and_summarize_cell_seg_data(
  csd_files,
  output_dir,
  study_dir = NULL,
  export_dir = NULL,
  require_include = FALSE,
  update_progress = NULL,
  col_select = NULL
)
}
\arguments{
\item{csd_files}{A list or vector of paths to cell seg data files.}

\item{output_dir}{Path to a directory where the results will be saved.}

\item{study_dir}{Optional path to a directory containing annotation files
for the images in the analysis. If provided, the files will be searched
for tagged ROIs. Cells in ROIs tagged with \verb{#ExcludeFromResults} will
be omitted from the consolidated data file.}

\item{export_dir}{If \code{study_dir} is provided, the path to an export folder
containing \code{binary_seg_maps} files and merged cell seg summary files
for the annotations referenced in \code{csd_files}.}

\item{require_include}{If \code{study_dir} is provided, should the result
include only cells contained in ROIs tagged with \verb{#IncludeInResults}?}

\item{update_progress}{Callback function which is called with progress.}

\item{col_select}{Column selection for \code{\link[phenoptr:read_cell_seg_data]{phenoptr::read_cell_seg_data()}}}
}
\value{
A single data frame containing consolidated data and columns for each
single phenotype, invisibly.
}
\description{
Consolidate several cell seg data files,
each with its own \code{Phenotype} column,
into a single file with separate columns for each phenotype.
}
\details{
Create a summary report for each source file and the consolidated data.

Write the consolidated data to \code{Consolidated_data.txt} in the output
directory. Write a per-field summary of the consolidated data to
\code{Consolidated_data_summary.txt} in the output directory. The summary
file will reflect any trimming of the cells and tissue area.

The
individual files must all have exactly the same
\verb{Sample Name} or \verb{Annotation ID} and \verb{Cell ID}
columns. \link{split_phenotypes} is called to split the \code{Phenotype} columns.

Optionally, filter the resulting cell table by regions tagged
\verb{#IncludeInResults} or \verb{#ExcludeFromResults} and update tissue category
areas to reflect the new regions.
}
