% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_phenotypes.R
\name{count_phenotypes}
\alias{count_phenotypes}
\title{Count phenotypes per slide and tissue category}
\usage{
count_phenotypes(csd, phenotypes, tissue_categories)
}
\arguments{
\item{csd}{Cell seg data to use. This should already have been filtered
for the slides or fields of interest.}

\item{phenotypes}{A named list of phenotype selectors. Results will be
reported in the same order as phenotypes are listed here.}

\item{tissue_categories}{A character vector of tissue category names
of interest.}
}
\value{
A data frame with columns for Slide ID, Tissue Category, cell
counts for each requested phenotype, and total cells
}
\description{
Count phenotypes per slide and tissue category
}
