% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_helpers.R
\name{count_within_summary}
\alias{count_within_summary}
\title{Summarize "count within" distances}
\usage{
count_within_summary(
  csd,
  radii,
  phenotypes = NULL,
  categories = NA,
  details_path = NULL,
  .by = "Slide ID",
  extra_cols = NULL,
  whole_slide = FALSE
)
}
\arguments{
\item{csd}{Cell seg data with \verb{Cell X Position},
\verb{Cell Y Position}, field name and \code{Phenotype} columns.}

\item{radii}{Vector of radii to search within.}

\item{phenotypes}{Optional list of phenotypes to include. If omitted,
will use \code{unique_phenotypes(csd)}. Counts are computed for all
pairs of phenotypes.}

\item{categories}{Optional list of tissue categories to compute within. If
omitted or \code{NA}, all cells will be included.}

\item{details_path}{If present, path to save tab-separated tables with
count within data for each cell.}

\item{.by}{Column to aggregate by}

\item{extra_cols}{The names of extra columns to include in the detailed
results.}

\item{whole_slide}{If \code{TRUE}, process all fields for each slide as one
data set; if \code{FALSE}, fields are processed individually.}
}
\value{
A data frame with summary statistics for each phenotype pair
in each Slide ID.
}
\description{
Computes summary "count within" statistics
for each Slide ID in \code{csd} and each pair of phenotypes in \code{phenotypes}.
See \code{\link[phenoptr:count_within]{phenoptr::count_within()}} for details of the counts and the
summary calculation.
}
\details{
If \code{details_path} is provided, this will save a table with one
row per cell and columns for each phenotype and radius giving
the count of cells of that type within that distance.

Note: If \code{categories} is \code{NA} or includes \code{NA}, the "All" tissue category in
the results will include all cells in \code{csd}. If \code{categories} does not include
\code{NA}, the "All" tissue category will include only cells in the provided
categories.
}
