% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_gadget_addin.R
\name{merge_addin}
\alias{merge_addin}
\title{Merge and summarize cell seg data files}
\usage{
merge_addin()
}
\description{
\code{merge_addin} opens a GUI that allows you to select multiple
cell seg data files. The selected files will be consolidated
to a single file which is saved to an output directory
along with summary reports for each file.
}
