% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_maps.R
\name{nearest_neighbor_map}
\alias{nearest_neighbor_map}
\title{Make a nearest neighbor map for a single field}
\usage{
nearest_neighbor_map(
  csd,
  field_name,
  view_number,
  export_path,
  phenos,
  color1,
  color2,
  show_as = c("from_to", "to_from", "mutual", "none"),
  dot_size = 3,
  add_logo = TRUE
)
}
\arguments{
\item{csd}{Cell seg data with distance columns}

\item{field_name}{Sample Name or Annotation ID to map}

\item{view_number}{Image index within the composite data}

\item{export_path}{Path to a directory containing composite and component
image files from inForm}

\item{phenos}{Named list of phenotype definitions. Must have length 2.}

\item{color1, color2}{Colors to draw the phenotype dots}

\item{show_as}{Which nearest neighbors should be shown?}

\item{dot_size}{Size of the dots used to show phenotypes}

\item{add_logo}{Show the Akoya logo in the image?}
}
\value{
Returns a \code{list} containing two items:
\describe{
\item{\code{plot}}{The plot, a \code{\link[ggplot2]{ggplot}} object.}
\item{\code{data}}{A \code{\link[tibble]{tibble}} containing the data
used to create the line
segments in the plot, or \code{NULL} if \code{show_as} is \code{"none"}.}
}
}
\description{
The phenotype definitions may be NA, in which case the base field will be
shown with any available phenotype.
}
