% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_helpers.R
\name{nearest_neighbor_summary}
\alias{nearest_neighbor_summary}
\title{Summarize nearest neighbor distances}
\usage{
nearest_neighbor_summary(csd, phenotypes = NULL, details_path = NULL)
}
\arguments{
\item{csd}{Cell seg data with \code{Cell X Position},
\code{Cell Y Position}, field name and \code{Phenotype} columns.}

\item{phenotypes}{Optional list of phenotypes to include. If omitted,
will use \code{unique_phenotypes(csd)}.}

\item{details_path}{If present, path to save a table with
nearest-neighbor data for each cell.}
}
\value{
A data frame with summary statistics for each phenotype pair
in each Slide ID.
}
\description{
Computes summary nearest neighbor statistics (mean, median, etc)
for each Slide ID in \code{csd} and each pair of phenotypes in \code{phenotypes}.
}
