% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{parse_phenotypes}
\alias{parse_phenotypes}
\title{Parse a vector of phenotype names to create selectors for
\code{\link[phenoptr:select_rows]{phenoptr::select_rows()}}.}
\usage{
parse_phenotypes(phenos)
}
\arguments{
\item{phenos}{A character vector of phenotype names.}
}
\value{
A named list of phenotype selectors
}
\description{
Parse a vector of phenotype names to create selectors for
\code{\link[phenoptr:select_rows]{phenoptr::select_rows()}}.
}
\section{Details}{

Each name must be either a single phenotype name (e.g. CD3+ or CD8-)
or two or more names separated by a slash (/). Additionally,
a name without a + or - and containing either "Total" or "All" will be
interpreted as meaning "All cells".
}

