% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_phenotypes.R
\name{split_phenotypes}
\alias{split_phenotypes}
\title{Split a phenotype column}
\usage{
split_phenotypes(csd)
}
\arguments{
\item{csd}{Cell seg data to use.}
}
\value{
A new data frame with the \code{Phenotype} column replaced with
individual columns per phenotype and the \code{Confidence} column removed.
}
\description{
A column containing multiple phenotypes is split
into multiple columns, one for each single phenotype.
}
\details{
Multiple phenotypes in the original column must be separated with "/".
The names of positive phenotypes must end in "+".
}
