% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{validate_phenotype_definitions}
\alias{validate_phenotype_definitions}
\title{Check that a user-specified phenotype definition
can be formed from available phenotypes.}
\usage{
validate_phenotype_definitions(pheno, available)
}
\arguments{
\item{pheno}{Text destription of a phenotype,
for \code{phenoptr::parse_phenotypes}.}

\item{available}{A character vector of available phenotypes}
}
\value{
An error message or empty string
}
\description{
Check that a user-specified phenotype definition
can be formed from available phenotypes.
}
