% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel_helpers.R
\name{write_percents_sheet}
\alias{write_percents_sheet}
\title{Write a cell percent table to an Excel workbook}
\usage{
write_percents_sheet(wb, percents, sheet_name = "Cell Percents",
  sheet_title = "Percentage of Total Cells")
}
\arguments{
\item{wb}{An openxlsx Workbook from \link[openxlsx:createWorkbook]{openxlsx::createWorkbook}}

\item{percents}{A data frame with columns for \code{Slide ID}, \code{Tissue Category},
and percent.}

\item{sheet_name}{Optional name for the worksheet.}

\item{sheet_title}{Optional title header for the table.}
}
\description{
Write a formatted cell percent table to a sheet in an Excel workbook.
}
\seealso{
Other output functions: \code{\link{write_counts_sheet}},
  \code{\link{write_density_sheet}},
  \code{\link{write_expression_sheet}},
  \code{\link{write_h_score_sheet}},
  \code{\link{write_nearest_neighbor_summary_sheet}},
  \code{\link{write_plot_sheet}},
  \code{\link{write_summary_sheet}}
}
\concept{output functions}
