% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel_helpers.R
\name{write_sheet}
\alias{write_sheet}
\title{Write a single sheet with formatting common to all sheets.}
\usage{
write_sheet(
  wb,
  d,
  sheet_name,
  sheet_title,
  header_col,
  keepNA = TRUE,
  addGrid = TRUE
)
}
\arguments{
\item{wb}{An openxlsx Workbook from \link[openxlsx:createWorkbook]{openxlsx::createWorkbook}}

\item{d}{A data frame to write.}

\item{sheet_name}{Name for the worksheet.}

\item{sheet_title}{Title header for the plot.}

\item{header_col}{Column number to start the \code{sheet_title}}

\item{keepNA}{If TRUE, NA values are written as #N/A; else they are blank.}

\item{addGrid}{If TRUE, grid lines (and page breaks) are added based
on the tissue categories in the data. If FALSE, no grid lines are added
and page breaks are added where needed.}
}
\description{
\itemize{
\item Bold, centered header
\item Bold column headers
\item Bold Slide ID column
\item Two-row page title
\item Page breaks as needed
}
}
